+++ src/Provider/AuthProviderFactory.php
@@ -8,23 +8,23 @@
  * @license http://opensource.org/licenses/gpl-license.php GNU Public License
  */
 
-namespace SocialAuther\Adapter;
+namespace SocialAuther\Provider;
 
 
-class AdapterFactory
+class AuthProviderFactory
 {
     /**
      * @var array mapping provider name => provider class for Factory
      */
     protected static $map = array(
         // provider name => provider class
-        AdapterBase::PROVIDER_FACEBOOK      => Facebook::class,
-        AdapterBase::PROVIDER_GOOGLE        => Google::class,
-        AdapterBase::PROVIDER_MAILRU        => Mailru::class,
-        AdapterBase::PROVIDER_ODNOKLASSNIKI => Odnoklassniki::class,
-        AdapterBase::PROVIDER_TWITTER       => Twitter::class,
-        AdapterBase::PROVIDER_VKONTAKTE     => Vkontakte::class,
-        AdapterBase::PROVIDER_YANDEX        => Yandex::class,
+        AuthProviderBase::PROVIDER_FACEBOOK      => Facebook::class,
+        AuthProviderBase::PROVIDER_GOOGLE        => Google::class,
+        AuthProviderBase::PROVIDER_MAILRU        => Mailru::class,
+        AuthProviderBase::PROVIDER_ODNOKLASSNIKI => Odnoklassniki::class,
+        AuthProviderBase::PROVIDER_TWITTER       => Twitter::class,
+        AuthProviderBase::PROVIDER_VKONTAKTE     => Vkontakte::class,
+        AuthProviderBase::PROVIDER_YANDEX        => Yandex::class,
     );
 
     public static function create($provider, $config)
