+++ src/Provider/Google.php
@@ -9,38 +9,38 @@
  * @license http://opensource.org/licenses/gpl-license.php GNU Public License
  */
 
-namespace SocialAuther\Adapter;
+namespace SocialAuther\Provider;
 
-class Mailru extends AdapterBase
+class Google extends AuthProviderBase
 {
     /**
      * {@inheritDoc}
      */
-    protected $provider = self::PROVIDER_MAILRU;
+    protected $provider = self::PROVIDER_GOOGLE;
 
     /**
      * {@inheritDoc}
      */
     protected $fieldsMap = array(
         // local property name => external property name
-        self::ATTRIBUTE_ID         => 'uid',
+        self::ATTRIBUTE_ID         => 'id',
         self::ATTRIBUTE_EMAIL      => 'email',
-        self::ATTRIBUTE_NAME       => 'nick',
+        self::ATTRIBUTE_NAME       => 'name',
         self::ATTRIBUTE_PAGE_URL   => 'link',
-        self::ATTRIBUTE_AVATAR_URL => 'pic_big',
-        self::ATTRIBUTE_BIRTHDAY   => 'birthday',
+        self::ATTRIBUTE_SEX        => 'gender',
+        self::ATTRIBUTE_AVATAR_URL => 'picture',
     );
 
     /**
-     * Get user sex or null if it is not set
+     * Get user birthday or null if it is not set
      *
      * @return string|null
      */
-    public function getSex()
+    public function getBirthday()
     {
-        if (isset($this->userInfo['sex'])) {
-            // gender is specified
-            return $this->userInfo['sex'] == 1 ? 'female' : 'male';
+        if (isset($this->userInfo['birthday'])) {
+            $this->userInfo['birthday'] = str_replace('0000', date('Y'), $this->userInfo['birthday']);
+            return date('d.m.Y', strtotime($this->userInfo['birthday']));
         }
 
         return null;
@@ -66,30 +66,24 @@ class Mailru extends AdapterBase
         );
 
         // Perform auth
-        $authInfo = $this->post('https://connect.mail.ru/oauth/token', $params);
+        $authInfo = $this->post('https://accounts.google.com/o/oauth2/token', $params);
         if (!isset($authInfo['access_token'])) {
             // something went wrong
             return false;
         }
 
         // Auth OK, can fetch additional info
-        $params = array(
-            'method'      => 'users.getInfo',
-            'secure'      => '1',
-            'app_id'      => $this->clientId,
-            'session_key' => $authInfo['access_token'],
-            'sig'         => md5("app_id={$this->clientId}method=users.getInfosecure=1session_key={$authInfo['access_token']}{$this->clientSecret}"),
-        );
+        $params['access_token'] = $authInfo['access_token'];
 
         // Fetch user info
-        $userInfo = $this->get('http://www.appsmail.ru/platform/api', $params);
-        if (!isset($userInfo[0][$this->fieldsMap[static::ATTRIBUTE_ID]])) {
+        $userInfo = $this->get('https://www.googleapis.com/oauth2/v1/userinfo', $params);
+        if (!isset($userInfo[$this->fieldsMap[static::ATTRIBUTE_ID]])) {
             // something went wrong
             return false;
         }
 
         // user info received
-        $this->userInfo = array_shift($userInfo);
+        $this->userInfo = $userInfo;
 
         return true;
     }
@@ -100,11 +94,12 @@ class Mailru extends AdapterBase
     public function getAuthUrlComponents()
     {
         return array(
-            'auth_url'    => 'https://connect.mail.ru/oauth/authorize',
+            'auth_url'    => 'https://accounts.google.com/o/oauth2/auth',
             'auth_params' => array(
-                'client_id'     => $this->clientId,
+                'redirect_uri'  => $this->redirectUri,
                 'response_type' => 'code',
-                'redirect_uri'  => $this->redirectUri
+                'client_id'     => $this->clientId,
+                'scope'         => 'https://www.googleapis.com/auth/userinfo.email https://www.googleapis.com/auth/userinfo.profile'
             )
         );
     }
