+++ src/Provider/Odnoklassniki.php
@@ -9,37 +9,54 @@
  * @license http://opensource.org/licenses/gpl-license.php GNU Public License
  */
 
-namespace SocialAuther\Adapter;
+namespace SocialAuther\Provider;
 
-class Facebook extends AdapterBase
+class Odnoklassniki extends AuthProviderBase
 {
+    /**
+     * Social Public Key
+     *
+     * @var string|null
+     */
+    protected $publicKey = null;
+
     /**
      * {@inheritDoc}
      */
-    protected $provider = self::PROVIDER_FACEBOOK;
+    protected $provider = self::PROVIDER_ODNOKLASSNIKI;
 
     /**
      * {@inheritDoc}
      */
     protected $fieldsMap = array(
         // local property name => external property name
-        self::ATTRIBUTE_ID       => 'id',
-        self::ATTRIBUTE_EMAIL    => 'email',
-        self::ATTRIBUTE_NAME     => 'name',
-        self::ATTRIBUTE_PAGE_URL => 'link',
-        self::ATTRIBUTE_SEX      => 'gender',
-        self::ATTRIBUTE_BIRTHDAY => 'birthday',
+        self::ATTRIBUTE_ID         => 'uid',
+        self::ATTRIBUTE_EMAIL      => 'email',
+        self::ATTRIBUTE_NAME       => 'name',
+        self::ATTRIBUTE_AVATAR_URL => 'pic_2',
+        self::ATTRIBUTE_SEX        => 'gender',
+        self::ATTRIBUTE_BIRTHDAY   => 'birthday',
     );
 
     /**
-     * Get url of user's avatar or null if it is not set
+     * {@inheritDoc}
+     */
+    protected $knownConfigParams = array(
+        'client_id',
+        'client_secret',
+        'redirect_uri',
+        'public_key',
+    );
+
+    /**
+     * Get user social id or null if it is not set
      *
      * @return string|null
      */
-    public function getAvatarUrl()
+    public function getPageUrl()
     {
-        if (isset($this->userInfo['username'])) {
-            return "http://graph.facebook.com/{$this->userInfo['username']}/picture?type=large";
+        if (isset($this->userInfo['uid'])) {
+            return 'http://www.odnoklassniki.ru/profile/' . $this->userInfo['uid'];
         }
 
         return null;
@@ -50,6 +67,11 @@ class Facebook extends AdapterBase
      */
     public function authenticate($params)
     {
+        if (!isset($this->publicKey)) {
+            // need to have own public key specified
+            return false;
+        }
+
         $params = $this->getAuthenticationParams($params);
         if (empty($params)) {
             // no required params provided
@@ -61,10 +83,11 @@ class Facebook extends AdapterBase
             'client_secret' => $this->clientSecret,
             'redirect_uri'  => $this->redirectUri,
             'code'          => $params['code'],
+            'grant_type'    => 'authorization_code',
         );
 
         // Perform auth
-        $authInfo = $this->post('https://graph.facebook.com/oauth/access_token', $params);
+        $authInfo = $this->post('http://api.odnoklassniki.ru/oauth/token.do', $params);
         if (!isset($authInfo['access_token'])) {
             // something went wrong
             return false;
@@ -72,11 +95,15 @@ class Facebook extends AdapterBase
 
         // Auth OK, can fetch additional info
         $params = array(
-            'access_token' => $authInfo['access_token']
+            'method'          => 'users.getCurrentUser',
+            'access_token'    => $authInfo['access_token'],
+            'application_key' => $this->publicKey,
+            'format'          => 'json',
+            'sig'             => md5("application_key={$this->publicKey}format=jsonmethod=users.getCurrentUser" . md5("{$authInfo['access_token']}{$this->clientSecret}")),
         );
 
         // Fetch user info
-        $userInfo = $this->get('https://graph.facebook.com/me', $params);
+        $userInfo = $this->get('http://api.odnoklassniki.ru/fb.do', $params);
         if (!isset($userInfo[$this->fieldsMap[static::ATTRIBUTE_ID]])) {
             // something went wrong
             return false;
@@ -94,12 +121,11 @@ class Facebook extends AdapterBase
     public function getAuthUrlComponents()
     {
         return array(
-            'auth_url'    => 'https://www.facebook.com/dialog/oauth',
+            'auth_url'    => 'http://www.odnoklassniki.ru/oauth/authorize',
             'auth_params' => array(
                 'client_id'     => $this->clientId,
-                'redirect_uri'  => $this->redirectUri,
                 'response_type' => 'code',
-                'scope'         => 'email,user_birthday'
+                'redirect_uri'  => $this->redirectUri
             )
         );
     }
