+++ src/Provider/Vkontakte.php
@@ -9,9 +9,9 @@
  * @license http://opensource.org/licenses/gpl-license.php GNU Public License
  */
 
-namespace SocialAuther\Adapter;
+namespace SocialAuther\Provider;
 
-class Vkontakte extends AdapterBase
+class Vkontakte extends AuthProviderBase
 {
     /**
      * {@inheritDoc}
