+++ src/Provider/Yandex.php
@@ -9,42 +9,27 @@
  * @license http://opensource.org/licenses/gpl-license.php GNU Public License
  */
 
-namespace SocialAuther\Adapter;
+namespace SocialAuther\Provider;
 
-class Google extends AdapterBase
+class Yandex extends AuthProviderBase
 {
     /**
      * {@inheritDoc}
      */
-    protected $provider = self::PROVIDER_GOOGLE;
+    protected $provider = self::PROVIDER_YANDEX;
 
     /**
      * {@inheritDoc}
      */
     protected $fieldsMap = array(
-        // local property name => external property name
         self::ATTRIBUTE_ID         => 'id',
-        self::ATTRIBUTE_EMAIL      => 'email',
-        self::ATTRIBUTE_NAME       => 'name',
+        self::ATTRIBUTE_EMAIL      => 'default_email',
+        self::ATTRIBUTE_NAME       => 'real_name',
         self::ATTRIBUTE_PAGE_URL   => 'link',
-        self::ATTRIBUTE_SEX        => 'gender',
-        self::ATTRIBUTE_AVATAR_URL => 'picture',
+        self::ATTRIBUTE_AVATAR_URL => 'default_avatar_id',
+        self::ATTRIBUTE_SEX        => 'sex',
+        self::ATTRIBUTE_BIRTHDAY   => 'birthday',
     );
-
-    /**
-     * Get user birthday or null if it is not set
-     *
-     * @return string|null
-     */
-    public function getBirthday()
-    {
-        if (isset($this->userInfo['birthday'])) {
-            $this->userInfo['birthday'] = str_replace('0000', date('Y'), $this->userInfo['birthday']);
-            return date('d.m.Y', strtotime($this->userInfo['birthday']));
-        }
-
-        return null;
-    }
 
     /**
      * {@inheritDoc}
@@ -60,23 +45,25 @@ class Google extends AdapterBase
         $params = array(
             'client_id'     => $this->clientId,
             'client_secret' => $this->clientSecret,
-            'redirect_uri'  => $this->redirectUri,
             'code'          => $params['code'],
             'grant_type'    => 'authorization_code',
         );
 
         // Perform auth
-        $authInfo = $this->post('https://accounts.google.com/o/oauth2/token', $params);
+        $authInfo = $this->post('https://oauth.yandex.ru/token', $params);
         if (!isset($authInfo['access_token'])) {
             // something went wrong
             return false;
         }
 
         // Auth OK, can fetch additional info
-        $params['access_token'] = $authInfo['access_token'];
+        $params = array(
+            'format'      => 'json',
+            'oauth_token' => $authInfo['access_token']
+        );
 
         // Fetch user info
-        $userInfo = $this->get('https://www.googleapis.com/oauth2/v1/userinfo', $params);
+        $userInfo = $this->get('https://login.yandex.ru/info', $params);
         if (!isset($userInfo[$this->fieldsMap[static::ATTRIBUTE_ID]])) {
             // something went wrong
             return false;
@@ -94,12 +81,11 @@ class Google extends AdapterBase
     public function getAuthUrlComponents()
     {
         return array(
-            'auth_url'    => 'https://accounts.google.com/o/oauth2/auth',
+            'auth_url'    => 'https://oauth.yandex.ru/authorize',
             'auth_params' => array(
-                'redirect_uri'  => $this->redirectUri,
                 'response_type' => 'code',
                 'client_id'     => $this->clientId,
-                'scope'         => 'https://www.googleapis.com/auth/userinfo.email https://www.googleapis.com/auth/userinfo.profile'
+                'display'       => 'popup'
             )
         );
     }
